backend/requirements.txt
Authlib==1.6.6
bleach==6.3.0
cryptography==44.0.3
email-validator==2.3.0
faker==40.1.2
fastapi==0.123.0
fhir.resources==8.2.0
httpx==0.28.1
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.10
openai==2.8.1
passlib[bcrypt]==1.7.4
pdfplumber==0.11.9
pgvector==0.4.2
pillow==12.0.0
psycopg[binary]==3.3.2
psycopg_pool==3.3.0
pydantic==2.11.7
pydantic-settings==2.12.0
pyotp==2.9.0
pytesseract==0.3.13
pytest==9.0.1
python-dotenv==1.1.1
python-multipart==0.0.20
qrcode==8.2
redis==7.1.0
slowapi==0.1.9
structlog==25.5.0
supabase==2.27.1
tenacity==8.5.0
uvicorn[standard]==0.35.0
